Refactor, common steps and dynamic paths
