+++ README.md
@@ -1,2 +1,4 @@
 # Cracking-Codes-with-Python
 Cracking Codes with Python
+
+poetry run python tests/base64/runTests.py
