+++ tests/base64/createOutputDir.py
@@ -1,15 +1,16 @@
-import os
 import sys
+import os
+
+current_directory = os.getcwd()
+steps_path = os.path.join(current_directory, 'steps')
+print(steps_path)
+sys.path.append(steps_path)
+
+import commonSteps
 
 script_directory = os.path.dirname(os.path.abspath(sys.argv[0]))
 
 # Define the path for the new directory
 folder_path = script_directory + "/output"
 
-# Create an empty directory if it doesn't exist
-if not os.path.exists(folder_path):
-    os.makedirs(folder_path)
-    print(f"Directory '{folder_path}' created successfully.")
-else:
-    print(f"Directory '{folder_path}' already exists.")
-
+commonSteps.createDir(folder_path)
