+++ tests/base64/decodeBase64ToFile.py
@@ -1,26 +1,14 @@
 import base64
 import sys
+import os
 
-def decode_base64_to_file(file_path):
-    try:
-        with open(file_path, "rb") as file:
-            file_content = file.read()
-            decoded_content = base64.b64decode(file_content)
-            return decoded_content
-    except FileNotFoundError:
-        print(f"File '{file_path}' not found.")
-        return None
-    except Exception as e:
-        print(f"An error occurred: {e}")
-        return None
+current_directory = os.getcwd()
+steps_path = os.path.join(current_directory, 'steps')
+print(steps_path)
+sys.path.append(steps_path)
 
-def write_decoded_data_to_file(file_path, decoded_data):
-    try:
-        with open(file_path, "wb") as output_file:
-            output_file.write(decoded_data)
-        print(f"Encoded content successfully written to {file_path}")
-    except Exception as e:
-        print(f"Error writing to file: {e}")
+import commonSteps
+
 
 if __name__ == "__main__":
     if len(sys.argv) != 3:
@@ -29,7 +17,7 @@ if __name__ == "__main__":
         input_file_path = sys.argv[1]
         output_file_path = sys.argv[2]
         
-        decoded_data = decode_base64_to_file(input_file_path)
+        decoded_data = commonSteps.decode_base64_to_file(input_file_path)
         
         if decoded_data:
-            write_decoded_data_to_file(output_file_path, decoded_data)
+            commonSteps.write_data_to_file(output_file_path, decoded_data)
